Refactor technician availability logic to static helper
Centralize technician availability checks in a new static AvailabilityLogic class. Move logic out of ClientService and update services to use the new helper. Add GetCandidateTechsForRequestAsync to repository for initial DB filtering, and apply detailed gap logic in code. Improve interface clarity and maintainability by separating DB and business logic.
