+++ el7erafe.Web/Core/DomainLayer/Contracts/ITechnicianAvailabilityRepository.cs
@@ -14,6 +14,13 @@ namespace DomainLayer.Contracts
         Task<bool> ExistsAsync(int id);
         Task<bool> ExistsForTechnicianAsync(int technicianId);
         Task<ICollection<string>> GetAvailableTechsForRequestAsync(int serviceId, int govId, WeekDay date, TimeOnly? from, TimeOnly? to, TimeOnly? minTime);
+        Task<List<Technician>> GetCandidateTechsForRequestAsync(
+    int serviceId,
+    int govId,
+    WeekDay date,
+    TimeOnly? requestedFrom,
+    TimeOnly? requestedTo,
+    TimeOnly? minimumStartTime);
         Task<int> DeleteAsync(int id);
         Task<int> DeleteByTechnicianIdAsync(int technicianId);
     }
