+++ el7erafe.Web/Core/ServiceAbstraction/ITechnicianAvailabilityService.cs
@@ -7,7 +7,14 @@ namespace ServiceAbstraction
     {
         Task<List<TechnicianAvailabilityResponseDto>> GetTechnicianAvailableTimeAsync(string technicianId);
 
-        Task<List<TechnicianAvailabilityResponseDto>> CreateScheduleAsync(string technicianId,List<AvailabilityBlockDto> blocks);
-        Task<List<string>> GetAvailableTechnicianByUserIdsAsync(int serviceId, int govId, DateOnly date, TimeOnly? from, TimeOnly? to, TimeOnly? minTime);
+        Task<List<TechnicianAvailabilityResponseDto>> CreateScheduleAsync(string technicianId, List<AvailabilityBlockDto> blocks);
+
+        Task<List<string>> GetAvailableTechnicianByUserIdsAsync(
+    int serviceId,
+    int govId,
+    DateOnly date,
+    TimeOnly? from,
+    TimeOnly? to,
+    TimeOnly? minTime);
     }
 }
